+++ NEWS.md
@@ -1,3 +1,9 @@
+- v0.2.0
+  - Add in parsenewick() to parse newick trees using Tokenize
+  - Rename iterator and filter functions to nodeiter(),
+     branchnamefilter(), etc and use function first ordering to allow
+     do block.
+  - Major bugfixes and addition of extensive testing of code.
 - v0.1.3
   - show() now working for trees again
   - Now provides implementation of getparent(), getancestors(), getchildren(), getdescendants()